requirements.txt
flask
tensorflow-cpu
opencv-python
numpy
pandas
pillow
scikit-learn
gunicorn
gdown